+++ internal/cli/build_amp_delegation.go
@@ -1,6 +1,0 @@
-package cli
-
-import _ "embed"
-
-//go:embed amp_delegation.ts
-var nativeAmpDelegation string